Refactor transaction services to Posting model
- TransactionService.CreateAsync/DeleteAsync rewritten to split into postings and recalc account balances
- DeleteAsync: soft-deletes each posting (DeletedAt); recalc balance for storage over all time, for category over current month
- Repo-layer models renamed to avoid ambiguity: PostingModel->PostingData, TransactionModel->TransactionData
- Money types in service models (request/response) set to decimal: CreateTransactionModel.Amount, endpoint Limitation/Value
- Financial analytics DTO models (CategoryComparativeAnalysis*, CategorySpendingItemModel, SpendingTrendLineNodeModel) use decimal
- ServiceModelMapper maps TransactionData/PostingData to service TransactionModel/PostingModel

+++ Service.Models/CreateModels/CreateTransactionEndpointModel.cs
@@ -18,7 +18,7 @@ public class CreateTransactionEndpointModel
 	/// <summary>
 	/// Ограничение трат
 	/// </summary>
-	public double? Limitation { get; set; }
+	public decimal? Limitation { get; set; }
 	
 	/// <summary>
 	/// Флаг указывающий, что конечная точка является денежным хранилищем
