+++ Service.Models/Models/FinancialAnalytics/CategoryComparativeAnalysisResult.cs
@@ -6,10 +6,10 @@ namespace Nasurino.SmartWallet.Service.Models.Models.FinancialAnalytics;
 public sealed class CategoryComparativeAnalysisResult
 {
     /// <summary>Общая сумма трат во втором периоде</summary>
-    public double TotalSecondPeriodSpending { get; set; }
+    public decimal TotalSecondPeriodSpending { get; set; }
 
     /// <summary>Общая сумма трат в первом периоде</summary>
-    public double TotalFirstPeriodSpending { get; set; }
+    public decimal TotalFirstPeriodSpending { get; set; }
 
     /// <summary>Сравнительные анализы по категориям</summary>
     public IReadOnlyCollection<CategoryComparativeAnalysisModel> CategoryComparativeAnalyses { get; set; }
